requirements-enterprise.txt
# Enterprise-level requirements for Seventh Day Sabbath Church Of Christ website
# Production-ready dependencies with specific versions for stability

# Core Django
Django>=5.0,<6.0
django-environ>=0.11.0
whitenoise>=6.6.0
Pillow>=10.0.0

# Forms and UI
django-crispy-forms>=2.0
crispy-tailwind>=0.5.0
django-widget-tweaks>=1.4.12

# Database
psycopg2-binary>=2.9.0
dj-database-url>=3.0.0

# Configuration
python-decouple>=3.8

# Security enhancements
django-ratelimit>=4.1.0
django-csp>=3.7
django-cors-headers>=4.3.0
django-security>=0.17.0

# Performance optimization
django-redis>=5.4.0
django-compressor>=4.4
django-imagekit>=4.1.0
django-storages>=1.14.0

# SEO and analytics
django-sitemap>=2.2
django-meta>=2.2.0
django-robots>=6.1

# Monitoring and logging
sentry-sdk[django]>=1.38.0
django-extensions>=3.2.3
django-debug-toolbar>=4.2.0

# API and integrations
requests>=2.31.0
celery>=5.3.0
redis>=5.0.0

# Email
django-ses>=3.5.0

# File handling
boto3>=1.29.0
django-cleanup>=8.0.0

# Development and testing (optional for production)
pytest-django>=4.7.0
coverage>=7.3.0
factory-boy>=3.3.0
faker>=20.1.0

# Additional enterprise features
django-health-check>=3.17.0
django-maintenance-mode>=0.21.1
django-admin-honeypot>=1.1.0
